feat: update models, controllers, and routers at cart product ordering flow

+++ src/models/cart.go
@@ -8,28 +8,29 @@ import (
 )
 
 type Cart struct {
-	Id              int        `db:"id" json:"id"`
-	OrderDetailId   int        `db:"orderDetailId" json:"orderDetailId" form:"orderDetailId"`
-	ProductName     *string    `db:"productName" json:"productName" form:"productName"`
-	ProductSize     *string    `db:"productSize" json:"productSize" form:"productSize"`
-	ProductVariant  *string    `db:"productVariant" json:"productVariant" form:"productVariant"`
-	Quantity        *int       `db:"quantity" json:"quantity" form:"quantity"`
-	AdditionalPrice *int       `db:"additionalPrice" json:"additionalPrice" form:"additionalPrice"`
-	Total           *int       `db:"total" json:"total" form:"total"`
-	CreatedAt       *time.Time `db:"createdAt" json:"createdAt"`
-	UpdatedAt       *time.Time `db:"updatedAt" json:"updatedAt"`
+	Id               int        `db:"id" json:"id"`
+	ProductId        *int       `db:"productId" json:"productId" form:"productId"`
+	ProductSizeId    *int       `db:"productSizeId" json:"productSizeId" form:"productSizeId"`
+	ProductVariantId *int       `db:"productVariantId" json:"productVariantId" form:"productVariantId"`
+	Quantity         *int       `db:"quantity" json:"quantity" form:"quantity"`
+	Total            *float64   `db:"total" json:"total" form:"total"`
+	UserId           *int       `db:"userId" json:"userId" form:"userId"`
+	CreatedAt        *time.Time `db:"createdAt" json:"createdAt"`
+	UpdatedAt        *time.Time `db:"updatedAt" json:"updatedAt"`
 }
 
 type CartInfo struct {
 	TotalPrice int
 }
 
-func FindAllCarts(limit int, offset int) (services.Info, error) {
+func FindAllCarts(userId int, limit int, offset int) (services.Info, error) {
+	fmtUserId := fmt.Sprintf(`%v`, userId)
 	sql := `SELECT * FROM "cart" 
+	WHERE "userId" = ` + fmtUserId + `
 	ORDER BY "id" ASC
 	LIMIT $1
 	OFFSET $2`
-	sqlCount := `SELECT COUNT(*) FROM "cart"`
+	sqlCount := `SELECT COUNT(*) FROM "cart" WHERE "userId" = ` + fmtUserId
 	result := services.Info{}
 	data := []Cart{}
 	db.Select(&data, sql, limit, offset)
@@ -40,25 +41,26 @@ func FindAllCarts(limit int, offset int) (services.Info, error) {
 	return result, err
 }
 
-func FindOneCart(id int) (Cart, error) {
-	sql := `SELECT * FROM "cart" WHERE "id"=$1`
+func FindOneCart(id int, userId int) (Cart, error) {
+	sql := `SELECT * FROM "cart" WHERE "id" = $1 AND "userId" = $2`
 	data := Cart{}
-	err := db.Get(&data, sql, id)
+	err := db.Get(&data, sql, id, userId)
 	return data, err
 }
 
-func TotalPrice() (CartInfo, error) {
-	sql := `SELECT SUM("total") FROM "cart"`
+func TotalPrice(userId int) (CartInfo, error) {
+	sql := `SELECT SUM("total") FROM "cart" 
+	WHERE "userId" = $1`
 	result := CartInfo{}
-	row := db.QueryRow(sql)
+	row := db.QueryRow(sql, userId)
 	err := row.Scan(&result.TotalPrice)
 	return result, err
 }
 
 func CreateCart(data Cart) (Cart, error) {
 	sql := `
-	INSERT INTO "cart" ("orderDetailId","productName","productSize","productVariant","quantity","additionalPrice","total") VALUES
-	(:orderDetailId, :productName,:productSize,:productVariant,:quantity,:additionalPrice,:total)
+	INSERT INTO "cart" ("productId","productSizeId","productVariantId","quantity","total","userId") VALUES
+	(:productId,:productSizeId,:productVariantId,:quantity,:total,:userId)
 	RETURNING *`
 
 	result := Cart{}
@@ -73,10 +75,12 @@ func CreateCart(data Cart) (Cart, error) {
 func UpdateCart(data Cart) (Cart, error) {
 	sql := `
 	UPDATE "cart" SET
-	"orderDetailId"=:orderDetailId
-	"quantity"=COALESCE(NULLIF(:quantity,0),"quantity"),
+	"productSizeId" = COALESCE(NULLIF(:productSizeId,0), "productSizeId"),
+	"productVariantId" = COALESCE(NULLIF(:productVariantId,0), "productVariantId"),
+	"quantity" = COALESCE(NULLIF(:quantity,0), "quantity"),
+	"total" = :total,
 	"updatedAt"=NOW()
-	WHERE id = :id
+	WHERE "id" = :id AND "userId" = :userId
 	RETURNING *
 	`
 	result := Cart{}
@@ -91,16 +95,18 @@ func UpdateCart(data Cart) (Cart, error) {
 	return result, err
 }
 
-func DeleteCart(id int) (Cart, error) {
-	sql := `DELETE FROM "cart" WHERE "id" = $1 RETURNING *`
+func DeleteCart(id int, userId int) (Cart, error) {
+	sql := `DELETE FROM "cart" 
+	WHERE "id" = $1 AND "userId" = $2
+	RETURNING *`
 	data := Cart{}
-	err := db.Get(&data, sql, id)
+	err := db.Get(&data, sql, id, userId)
 	return data, err
 }
 
-func DeleteAllCart() (Cart, error) {
-	sql := `DELETE FROM "cart" RETURNING *`
+func DeleteAllCart(userId int) (Cart, error) {
+	sql := `DELETE FROM "cart" WHERE "userId" = $1 RETURNING *`
 	data := Cart{}
-	err := db.Get(&data, sql)
+	err := db.Get(&data, sql, userId)
 	return data, err
 }
